NE-1 Remove quests relating to the Bane'thil Barrow Den. (#3)
The exp/time ratio for the quests in the Ban'ethil Barrow Den area are not good enough to be worth pursuing as part of a speedrun. This commit will also remove elite quests that are time-consuming to solo, or are a bottle-neck on new server launch day (e.g., Mist).